+++ cmd/frontend/graphqlbackend/search.go
@@ -69,7 +69,9 @@ func (r *schemaResolver) Search(args *struct {
 		return newSearcherResolver(query2)
 	}
 
-	go addQueryToSearchesTable(args.Query)
+	// TODO(ijt): remove this potential goroutine leak.
+	go r.addQueryToSearchesTable(args.Query)
+
 	query, err := query.ParseAndCheck(args.Query)
 	if err != nil {
 		log15.Debug("graphql search failed to parse", "query", args.Query, "error", err)
@@ -80,13 +82,17 @@ func (r *schemaResolver) Search(args *struct {
 	}, nil
 }
 
-func addQueryToSearchesTable(q string) {
+func (r *schemaResolver) addQueryToSearchesTable(q string) {
+	rs := r.recentSearches
 	ctx, cancel := context.WithTimeout(context.Background(), time.Second)
 	defer cancel()
-	if err := db.RecentSearches.Add(ctx, q); err != nil {
+	if rs == nil {
+		return
+	}
+	if err := rs.Log(ctx, q); err != nil {
 		log15.Error("adding query to searches table", "error", err)
 	}
-	if err := db.RecentSearches.DeleteExcessRows(ctx, 1e5); err != nil {
+	if err := rs.Cleanup(ctx, 1e5); err != nil {
 		log15.Error("deleting excess rows from searches table", "error", err)
 	}
 }